Changing the way installer tars are handled.
Instead of guessing where the components json is located, leave that environment
variable unset and let the installer use its baked in location instead. However,
if the tar location is a local path, at least make an attempt to find the
components. (This would happen if someone unzipped the full repo tar and then
tried to point to the installer tar within that.)